requirements.txt
allure-python-commons
pydantic
requests
pytest
